style: ruff auto-format on compact.py
